Refactoring how http exceptions is handled.
