+++ data/src/main/java/br/stone/mobiletraining/samilasantos/data/service/randomFact/HandleRandomFactHttpExceptions.kt
@@ -1,30 +1,29 @@
-package br.stone.mobiletraining.samilasantos.data.service.common
+package br.stone.mobiletraining.samilasantos.data.service.randomFact
 
+import br.stone.mobiletraining.samilasantos.domain.common.Fact
 import br.stone.mobiletraining.samilasantos.domain.common.IntegrationExceptions
-import br.stone.mobiletraining.samilasantos.domain.searchFact.SearchFactExceptions
 import io.reactivex.Single
 import io.reactivex.SingleSource
 import io.reactivex.SingleTransformer
 import retrofit2.adapter.rxjava2.HttpException
 
-class HandleHttpErrors<T> : SingleTransformer<T, T> {
+class HandleRandomFactHttpExceptions : SingleTransformer<Fact, Fact> {
 
-    override fun apply(upstream: Single<T>): SingleSource<T> {
+    override fun apply(upstream: Single<Fact>): SingleSource<Fact> {
         return upstream.onErrorResumeNext(this::handleIfRestError)
     }
 
-    private fun handleIfRestError(incoming: Throwable): Single<T> =
+    private fun handleIfRestError(incoming: Throwable): Single<Fact> =
         if (incoming is HttpException) toInfrastructureError(incoming)
         else Single.error(incoming)
 
-    private fun toInfrastructureError(restError: HttpException): Single<T> {
+    private fun toInfrastructureError(restError: HttpException): Single<Fact> {
         val infraError = mapErrorWith(restError.code())
         return Single.error(infraError)
     }
 
     private fun mapErrorWith(code: Int) = when (code) {
         404 -> IntegrationExceptions.InfoNotFound
-        in 400..499 -> SearchFactExceptions.MalFormedQuery
         in 500..511 -> IntegrationExceptions.UnavailableProvider
         else -> IntegrationExceptions.InfoNotFound
     }
